# web/frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchEnvelope": {
        "additionalProperties": false,
        "properties": {
          "batch": {
            "additionalProperties": true,
            "title": "Batch",
            "type": "object"
          }
        },
        "required": [
          "batch"
        ],
        "title": "BatchEnvelope",
        "type": "object"
      },
...
      "BatchListEnvelope": {
        "additionalProperties": false,
        "properties": {
          "batches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Batches",
            "type": "array"
          }
        },
        "required": [
          "batches"
        ],
        "title": "BatchListEnvelope",
        "type": "object"
      },
      "BatchPreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "test_set_id": {
            "minLength": 1,
            "title": "Test Set Id",
            "type": "string"
          }
        },
        "required": [
          "test_set_id"
        ],
        "title": "BatchPreviewRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EvalFlow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/batch-runs": {
      "get": {
        "operationId": "list_batches_api_batch_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Batches",
        "tags": [
          "batch-runs"
        ]
      },
      "post": {
        "operationId": "create_batch_api_batch_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch",
        "tags": [
          "batch-runs"
        ]
      }
    },
    "/api/batch-runs/preview": {
      "post": {
        "operationId": "preview_batch_input_api_batch_runs_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Batch Input",
        "tags": [
          "batch-runs"
        ]
      }
    },
    "/api/batch-runs/{batch_id}": {
      "delete": {
        "operationId": "delete_batch_api_batch_runs__batch_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Batch",
        "tags": [
          "batch-runs"
        ]
      },
      "get": {
        "operationId": "get_batch_api_batch_runs__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "round_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Round Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch",
        "tags": [
          "batch-runs"
        ]
      },
      "put": {
        "operationId": "update_batch_api_batch_runs__batch_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Batch",
        "tags": [
          "batch-runs"
        ]
      }
    },
    "/api/batch-runs/{batch_id}/cancel": {
      "post": {
        "operationId": "cancel_batch_api_batch_runs__batch_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Batch",
        "tags": [
          "batch-runs"
        ]
      }
    },
    "/api/batch-runs/{batch_id}/cases": {
      "get": {
        "operationId": "list_batch_cases_api_batch_runs__batch_id__cases_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "round_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Round Id"
            }
          },
          {
            "in": "query",
            "name": "result",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Pass",
                    "Failed",
                    "Error"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Running",
                    "Pending"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCaseListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batch Cases",
        "tags": [
          "batch-runs"
        ]
      }
    },
    "/api/batch-runs/{batch_id}/cases/{case_run_id}": {
      "get": {
        "operationId": "get_batch_case_api_batch_runs__batch_id__cases__case_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_run_id",
            "required": true,
            "schema": {
              "title": "Case Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "round_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Round Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Batch Case Api Batch Runs  Batch Id  Cases  Case Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch Case",
        "tags": [
          "batch-runs"
        ]
      }
    },
    "/api/batch-runs/{batch_id}/cases/{case_run_id}/cancel": {
      "post": {
        "operationId": "cancel_batch_case_api_batch_runs__batch_id__cases__case_run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_run_id",
            "required": true,
            "schema": {
              "title": "Case Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Batch Case Api Batch Runs  Batch Id  Cases  Case Run Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Batch Case",
        "tags": [
          "batch-runs"
        ]
      }
    },
...
    "/api/batch-runs/{batch_id}/history": {
      "get": {
        "operationId": "list_batch_history_api_batch_runs__batch_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchHistoryListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batch History",
        "tags": [
          "batch-runs"
        ]
      }
    },
...
  }
}